# Improve contour performance in WCSAxes by 10-1000x