@larrybradley 2018-06-20
@astrofrog When will the new contour benchmarks be run?  I don't see them yet.  I assume this PR shouldn't be merged until the benchmarks have been run at least once with the old code?