@astrofrog 2018-06-20
@larrybradley - thanks for the review! I'll check today why the benchmarks haven't run yet.