@larrybradley 2018-06-20
@astrofrog Let me know when the benchmarks have been run and I'll merge this.