@astrofrog 2018-06-20
@larrybradley - they have now run:

http://www.astropy.org/astropy-benchmarks/#visualization.wcsaxes.time_contour_with_transform

Thanks!